Revert "feat: Support MXFP4 quantized dense models on AMD CDNA2/CDNA3 GPUs (#19143)"
This reverts commit 2cc52d83263281fbd8b9c3a33b9455df0c34c232.
