+++ python/pyproject_other.toml
@@ -93,7 +93,7 @@ tracing = [
 srt_hip = [
   "sglang[runtime_common]",
   "torch",
-  "petit_kernel==0.0.3",
+  "petit_kernel==0.0.2",
   "wave-lang==3.8.2",
 ]
