+++ python/sglang/srt/configs/model_config.py
@@ -946,7 +946,6 @@ class ModelConfig:
             "fbgemm_fp8",
             "w8a8_fp8",
             "petit_nvfp4",
-            "petit_mxfp4",
             "quark",
             "mxfp4",
             "auto-round",
@@ -971,7 +970,6 @@ class ModelConfig:
             "qoq",
             "w4afp8",
             "petit_nvfp4",
-            "petit_mxfp4",
             "quark",
             "modelslim",
         ]
@@ -980,7 +978,6 @@ class ModelConfig:
             "modelopt_fp4": ["modelopt"],
             "modelopt_mixed": ["modelopt"],
             "petit_nvfp4": ["modelopt"],
-            "petit_mxfp4": ["mxfp4", "quark"],
             "w8a8_int8": ["compressed-tensors", "compressed_tensors"],
             "w8a8_fp8": ["compressed-tensors", "compressed_tensors"],
         }
