+++ python/sglang/srt/server_args.py
@@ -112,7 +112,6 @@ QUANTIZATION_CHOICES = [
     "modelopt_fp4",
     "modelopt_mixed",
     "petit_nvfp4",
-    "petit_mxfp4",
     "w8a8_int8",
     "w8a8_fp8",
     "moe_wna16",
